Set up separate page for visual Flashbar tests
